+++ CHANGELOG.md
@@ -4,3 +4,5 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+- Implement quasi-binary decoder.